Fix ads badges conflict with player badges
